% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDC.R
\name{SSL}
\alias{SSL}
\title{Main function for analyzing the effect of an exposure on a continuous outcome}
\usage{
SSL(nScans = 20000, burn = 10000, thin = 10, y, x, z, lambda1 = 0.1,
  thetaA = 1, thetaB = 0.2 * dim(x)[2], lambda0 = "EM")
}
\arguments{
\item{nScans}{The number of MCMC scans to run}

\item{burn}{The number of MCMC scans that will be dropped as a burn-in}

\item{thin}{This number represents how many iterations between each scan 
that is kept}

\item{y}{The outcome to be analyzed}

\item{x}{An n by p matrix of potential confounders}

\item{z}{The treatment whose causal effect is to be estimated}

\item{lambda1}{A numeric value for lambda1}

\item{thetaA}{The first parameter of the beta prior on the overall sparsity level}

\item{thetaB}{The second parameter of the beta prior on the overall sparsity level}

\item{lambda0}{Either a numeric value to be used for the value of lambda0
or "EM" is specified to indicate that it will be estimated
via empirical Bayes}
}
\value{
An list of values that contain the treatment effect, confidence interval for the 
        treatment effect, as well as the corresponding quantities for the regression parameters
        for the confounders.
}
\description{
This function will take in the observed data and estimate a treatment effect. y,x, and z must
all be supplied, though all other parameters have pre-set values the user can proceed with unless
they wish to change the prior specification. We recommend using the EM option to estimate lambda0
as results can be very sensitive to this parameter and choosing it a-priori is a difficult task.
}
\examples{

put example run here
}

