% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDC.R
\name{SSLhetero}
\alias{SSLhetero}
\title{Estimate causal effect allowing for heterogeneous treatment effects}
\usage{
SSLhetero(nScans = 20000, burn = 10000, thin = 10, y, x, z,
  lambda1 = 0.1, thetaA = 1, thetaB = 0.2 * dim(x)[2], lambda0 = "EM",
  weight = NULL, kMax = 20, EMiterMax = 300)
}
\arguments{
\item{nScans}{The number of MCMC scans to run}

\item{burn}{The number of MCMC scans that will be dropped as a burn-in}

\item{thin}{This number represents how many iterations between each scan 
that is kept}

\item{y}{The outcome to be analyzed}

\item{x}{An n by p matrix of potential confounders}

\item{z}{The treatment whose causal effect is to be estimated}

\item{lambda1}{A numeric value for lambda1}

\item{thetaA}{The first parameter of the beta prior on the overall sparsity level}

\item{thetaB}{The second parameter of the beta prior on the overall sparsity level}

\item{lambda0}{Either a numeric value to be used for the value of lambda0
or "EM" is specified to indicate that it will be estimated
via empirical Bayes}

\item{weight}{The weight used to prioritize variables associated with treatment. This
parameter has a default of NULL, and should only be used when lambdo0
is provided instead of estimated using empirical Bayes}

\item{kMax}{The maximum number of covariates to be prioritized due to association with treatment}

\item{EMiterMax}{The maximum number of iterations to update EM algorithm. The algorithm is updated and
checked for convergence every 50th MCMC scan. We recommend a high value such as 300 or
500 for this parameter to ensure convergence, though the program will stop well short
of 300 in most applications if it has converged.}
}
\value{
An list of values that contain the treatment effect, confidence interval for the 
        treatment effect, and full posterior draws for the treatment effect.
}
\description{
This function will take in the observed data and estimate a treatment effect allowing for 
heterogeneous treatment effects. y,x, and z must
all be supplied, though all other parameters have pre-set values the user can proceed with unless
they wish to change the prior specification. We recommend using the EM option to estimate lambda0
as results can be very sensitive to this parameter and choosing it a-priori is a difficult task.
Note, however, that the EM option will take longer as it requires running multiple MCMCs instead
of just one
}
\examples{

## p can be larger than n, but we keep the number of covariates small here
## just for illustration so that the code will finish faster
## n=200 and p=200 takes ~5 minutes to run for example
n = 200
p = 20
x = matrix(rnorm(n*p), n, p)
z = rbinom(n, 1, p=pnorm(0.5 + 0.7*x[,1] + 0.3*x[,2]))
y = rnorm(n, mean=z + 0.3*x[,1] + 0.6*x[,2] + 0.5*x[,3] + 0.5*z*x[,1], sd=1)

ssl = SSLhetero(y=y, z=z, x=x, nScans=3000, burn=1000, thin=2)
## Output treatment effect and credible interval
print(ssl$TreatEffect)
print(ssl$TreatEffectCI)
}

